requirements.txt
sentence-transformers==2.2.2
pdfminer.six==20221105
torch>=1.12